+++ packages/stream_chat_flutter/lib/src/channel_list_header.dart
@@ -185,7 +185,7 @@ class ChannelListHeader extends StatelessWidget implements PreferredSizeWidget {
                       ),
                       onPressed: onNewChatButtonTap,
                     ),
-                  )
+                  ),
                 ],
             title: Column(
               children: [