+++ packages/stream_chat_flutter/lib/src/message_list_view.dart
@@ -60,6 +60,7 @@ typedef OnMessageTap = void Function(Message);
 typedef ReplyTapCallback = void Function(Message);
 
 /// Class for message details
+// ignore: prefer-match-file-name
 class MessageDetails {
   /// Constructor for creating [MessageDetails]
   MessageDetails(
@@ -358,8 +359,9 @@ class _MessageListViewState extends State<MessageListView> {
                   child: Text(
                     context.translations.emptyChatMessagesText,
                     style: _streamTheme.textTheme.footnote.copyWith(
-                        color: _streamTheme.colorTheme.textHighEmphasis
-                            .withOpacity(.5)),
+                      color: _streamTheme.colorTheme.textHighEmphasis
+                          .withOpacity(0.5),
+                    ),
                   ),
                 ),
         messageListBuilder: widget.messageListBuilder ??
@@ -371,8 +373,9 @@ class _MessageListViewState extends State<MessageListView> {
                   child: Text(
                     context.translations.genericErrorText,
                     style: _streamTheme.textTheme.footnote.copyWith(
-                        color: _streamTheme.colorTheme.textHighEmphasis
-                            .withOpacity(.5)),
+                      color: _streamTheme.colorTheme.textHighEmphasis
+                          .withOpacity(0.5),
+                    ),
                   ),
                 ),
       );
@@ -383,7 +386,7 @@ class _MessageListViewState extends State<MessageListView> {
 
     if (_messageListLength != null) {
       if (_bottomPaginationActive || (_inBetweenList && _upToDate)) {
-        if (_itemPositionListener.itemPositions.value.isNotEmpty == true) {
+        if (_itemPositionListener.itemPositions.value.isNotEmpty) {
           final first = _itemPositionListener.itemPositions.value.first;
           final diff = newMessagesListLength - _messageListLength!;
           if (diff > 0) {
@@ -635,10 +638,14 @@ class _MessageListViewState extends State<MessageListView> {
     );
 
     final backgroundColor = MessageListViewTheme.of(context).backgroundColor;
+    final backgroundImage = MessageListViewTheme.of(context).backgroundImage;
 
-    if (backgroundColor != null) {
-      return ColoredBox(
-        color: backgroundColor,
+    if (backgroundColor != null || backgroundImage != null) {
+      return Container(
+        decoration: BoxDecoration(
+          color: backgroundColor,
+          image: backgroundImage,
+        ),
         child: child,
       );
     }
@@ -674,7 +681,8 @@ class _MessageListViewState extends State<MessageListView> {
         child: BetterStreamBuilder<Iterable<ItemPosition>>(
           initialData: _itemPositionListener.itemPositions.value,
           stream: _valueListenableToStreamAdapter(
-              _itemPositionListener.itemPositions),
+            _itemPositionListener.itemPositions,
+          ),
           comparator: (a, b) {
             if (a == null || b == null) {
               return false;
@@ -961,7 +969,7 @@ class _MessageListViewState extends State<MessageListView> {
     }
 
     final userId = StreamChat.of(context).currentUser!.id;
-    final isMyMessage = message.user!.id == userId;
+    final isMyMessage = message.user?.id == userId;
     final nextMessage = index - 1 >= 0 ? messages[index - 1] : null;
     final isNextUserSame =
         nextMessage != null && message.user!.id == nextMessage.user!.id;
@@ -984,7 +992,7 @@ class _MessageListViewState extends State<MessageListView> {
 
     final allRead = readList.length >= (channel.memberCount ?? 0) - 1;
     final hasFileAttachment =
-        message.attachments.any((it) => it.type == 'file') == true;
+        message.attachments.any((it) => it.type == 'file');
 
     final isThreadMessage =
         message.parentId != null && message.showInChannel == true;
@@ -1016,7 +1024,7 @@ class _MessageListViewState extends State<MessageListView> {
     final isOnlyEmoji = message.text?.isOnlyEmoji ?? false;
 
     final hasUrlAttachment =
-        message.attachments.any((it) => it.titleLink != null) == true;
+        message.attachments.any((it) => it.titleLink != null);
 
     final borderSide =
         isOnlyEmoji || hasUrlAttachment || (isMyMessage && !hasFileAttachment)
@@ -1259,10 +1267,11 @@ class _MessageListViewState extends State<MessageListView> {
     if (widget.onThreadTap != null) {
       _onThreadTap = (Message message) {
         widget.onThreadTap!(
-            message,
-            widget.threadBuilder != null
-                ? widget.threadBuilder!(context, message)
-                : null);
+          message,
+          widget.threadBuilder != null
+              ? widget.threadBuilder!(context, message)
+              : null,
+        );
       };
     } else if (widget.threadBuilder != null) {
       _onThreadTap = (Message message) {
@@ -1271,7 +1280,8 @@ class _MessageListViewState extends State<MessageListView> {
           MaterialPageRoute(
             builder: (_) => BetterStreamBuilder<Message>(
               stream: streamChannel!.channel.state!.messagesStream.map(
-                  (messages) => messages.firstWhere((m) => m.id == message.id)),
+                (messages) => messages.firstWhere((m) => m.id == message.id),
+              ),
               initialData: message,
               builder: (_, data) => StreamChannel(
                 channel: streamChannel!.channel,
@@ -1320,7 +1330,7 @@ class _LoadingIndicator extends StatelessWidget {
       stream: stream,
       initialData: false,
       errorBuilder: (context, error) => Container(
-        color: streamTheme.colorTheme.accentError.withOpacity(.2),
+        color: streamTheme.colorTheme.accentError.withOpacity(0.2),
         child: Center(
           child: Text(context.translations.loadingMessagesError),
         ),