+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -354,8 +354,10 @@ class StreamChatLocalizationsFr extends GlobalStreamChatLocalizations {
   String get youText => 'Vous';
 
   @override
-  String galleryPaginationText(
-          {required int currentPage, required int totalPages}) =>
+  String galleryPaginationText({
+    required int currentPage,
+    required int totalPages,
+  }) =>
       '${currentPage + 1} de $totalPages';
 
   @override