+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fixed message highlight animation alignment in `MessageListView`
 
+✅ Added
+
+- `MessageListViewThemeData` now accepts a `DecorationImage` as a background image for `MessageListView`.
+
 ## 3.1.1
 
 - Updated `stream_chat_flutter_core` dependency to [`3.1.1`](https://pub.dev/packages/stream_chat_flutter_core/changelog).